Vision: Fixed perspective constant usage. #259

+++ vision/src/Perspective.cpp
@@ -10,16 +10,17 @@
 #include "Perspective.hpp"
 
 #include "cameraConstants.hpp"
+#include "tuum_platform.hpp"
 
 
 namespace rtx { namespace Vision { namespace Perspective {
 
-  std::vector<double> A = {rtx::gC.getInt("Perspective.First.A"),
-                           rtx::gC.getInt("Perspective.Second.A")};
-  std::vector<double> B = {rtx::gC.getInt("Perspective.First.B"),
-                           rtx::gC.getInt("Perspective.Second.B")};
-  std::vector<double> C = {rtx::gC.getInt("Perspective.First.C"),
-                           rtx::gC.getInt("Perspective.Second.C")};
+  std::vector<double> A = {(double) rtx::gC.getInt("Perspective.First.A"),
+                           (double) rtx::gC.getInt("Perspective.Second.A")};
+  std::vector<double> B = {(double) rtx::gC.getInt("Perspective.First.B"),
+                           (double) rtx::gC.getInt("Perspective.Second.B")};
+  std::vector<double> C = {(double) rtx::gC.getInt("Perspective.First.C"),
+                           (double) rtx::gC.getInt("Perspective.Second.C")};
 
   std::pair<double, double> virtualToReal(const unsigned int &x, const unsigned int &y, const unsigned int &cameraID) {
     // ActualDistance = A + B / PixelVerticalCoord
